CU-8hvhw6 - organize implement structure for paymnent and subscription module
