+++ src/main/java/org/elastos/hive/BackupSubscription.java
@@ -2,15 +2,15 @@ package org.elastos.hive;
 
 import org.elastos.hive.exception.HiveException;
 import org.elastos.hive.exception.UnsupportedMethodException;
-import org.elastos.hive.vault.payment.Order;
-import org.elastos.hive.vault.payment.PricingPlan;
-import org.elastos.hive.vault.payment.Receipt;
+import org.elastos.hive.subscription.payment.Order;
+import org.elastos.hive.subscription.payment.PaymentServiceRender;
+import org.elastos.hive.subscription.payment.PricingPlan;
+import org.elastos.hive.subscription.payment.Receipt;
 import org.elastos.hive.service.PaymentService;
 import org.elastos.hive.service.SubscriptionService;
 import org.elastos.hive.vault.ExceptionConvertor;
-import org.elastos.hive.vault.PaymentServiceRender;
-import org.elastos.hive.vault.SubscriptionServiceRender;
-import org.elastos.hive.vault.subscription.VaultInfoResponseBody;
+import org.elastos.hive.subscription.SubscriptionServiceRender;
+import org.elastos.hive.subscription.VaultInfoResponseBody;
 
 import java.io.IOException;
 import java.util.Collections;
