+++ src/main/java/org/elastos/hive/VaultSubscription.java
@@ -2,9 +2,10 @@ package org.elastos.hive;
 
 import org.elastos.hive.exception.HiveException;
 import org.elastos.hive.exception.UnsupportedMethodException;
-import org.elastos.hive.vault.payment.Order;
-import org.elastos.hive.vault.payment.PricingPlan;
-import org.elastos.hive.vault.payment.Receipt;
+import org.elastos.hive.subscription.payment.Order;
+import org.elastos.hive.subscription.payment.PaymentServiceRender;
+import org.elastos.hive.subscription.payment.PricingPlan;
+import org.elastos.hive.subscription.payment.Receipt;
 import org.elastos.hive.service.PaymentService;
 import org.elastos.hive.service.SubscriptionService;
 
@@ -15,9 +16,8 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
 
 import org.elastos.hive.vault.ExceptionConvertor;
-import org.elastos.hive.vault.PaymentServiceRender;
-import org.elastos.hive.vault.SubscriptionServiceRender;
-import org.elastos.hive.vault.subscription.VaultInfoResponseBody;
+import org.elastos.hive.subscription.SubscriptionServiceRender;
+import org.elastos.hive.subscription.VaultInfoResponseBody;
 
 public class VaultSubscription extends ServiceEndpoint
 	implements SubscriptionService<Vault.PropertySet>, PaymentService, ExceptionConvertor {
