+++ src/main/java/org/elastos/hive/subscription/SubscriptionController.java
@@ -1,4 +1,4 @@
-package org.elastos.hive.vault.subscription;
+package org.elastos.hive.subscription;
 
 import org.elastos.hive.ServiceEndpoint;
 import org.elastos.hive.connection.HiveResponseBody;
