+++ src/main/java/org/elastos/hive/subscription/payment/OrderInfoResponseBody.java
@@ -1,8 +1,7 @@
-package org.elastos.hive.vault.payment;
+package org.elastos.hive.subscription.payment;
 
 import com.google.gson.annotations.SerializedName;
 import org.elastos.hive.connection.HiveResponseBody;
-import org.elastos.hive.vault.payment.Order;
 
 class OrderInfoResponseBody extends HiveResponseBody {
     @SerializedName("order_info")
