+++ src/main/java/org/elastos/hive/subscription/payment/PaymentAPI.java
@@ -1,4 +1,4 @@
-package org.elastos.hive.vault.payment;
+package org.elastos.hive.subscription.payment;
 
 import org.elastos.hive.connection.HiveResponseBody;
 import retrofit2.Call;
