+++ src/main/java/org/elastos/hive/subscription/payment/Receipt.java
@@ -1,4 +1,4 @@
-package org.elastos.hive.vault.payment;
+package org.elastos.hive.subscription.payment;
 
 public class Receipt {
 	private String receiptId;
